Add files via upload
EDCRGTCcode is a package to Mathematica. 2005.07833.pdf is the article and I am trying to reproduce equations (5), (6) and (8). gbonnet variation alex is variation to get equations of motion. gbonnet alex is an attempt to solve eom in the nbh of x=0 and x=1. holographic renormalization.pdf is a proposed literature to get equation (8). Applied holography is some literature that could come in handy.